backend/requirements.txt
fastapi>=0.95.0
uvicorn[standard]>=0.22.0
sqlalchemy>=2.0.0
python-dotenv>=1.0.0
nltk>=3.8.0
faiss-cpu>=1.7.4
numpy>=1.23.0
google-generativeai>=0.1.0
python-docx>=0.8.11
PyPDF2>=3.0.0
requests>=2.28.0
